future2/requirements.txt
aiogram>=3.0.0
python-dotenv>=0.19.0
pandas>=1.3.0
numpy>=1.21.0
sentence-transformers>=2.2.0
qdrant-client>=1.1.0
pymongo>=4.0.0
beautifulsoup4>=4.9.3
emoji>=1.7.0
requests>=2.31.0
langchain>=0.1.0
langchain-community>=0.0.10
langchain-core>=0.1.10
langchain-deepseek>=0.1.0
langchain-openai>=0.0.2
google-generativeai==0.8.5
google-ai-generativelanguage==0.6.15
langchain-google-genai==0.0.5
langchain-ollama>=0.0.1
openai>=1.0.0
tiktoken>=0.5.0 
apscheduler==3.10.4 